+++ jsond/balance.go
@@ -32,8 +32,8 @@ func (b *Balance) UnmarshalJSON(bytes []byte) error {
 	)
 
 	b.Symbol = symbol
-	b.Available = util.IfOrElse(len(available) > 0, func() float64 { return util.MustFloat64(available) }, float64(0))
-	b.InOrder = util.IfOrElse(len(inOrder) > 0, func() float64 { return util.MustFloat64(inOrder) }, float64(0))
+	b.Available = util.IfOrElse(len(available) > 0, func() float64 { return util.MustFloat64(available) }, 0)
+	b.InOrder = util.IfOrElse(len(inOrder) > 0, func() float64 { return util.MustFloat64(inOrder) }, 0)
 
 	return nil
 }
